refactor: less exposure for perform context
